fix(ai): fix LM Studio model list deserialization crash and int→long for sizes
/api/v1/models returns { "models": [...] }, not a bare array; deserializing
as List<ModelInfo> threw a JSON exception. Introduced LmStudioModelResponse
wrapper. Also widened TotalSizeBytes, SizeBytes, FileSizeBytes to long to
handle models larger than ~2 GB without overflow.

+++ src/TombLauncher.Core/Dtos/AiModelMetadata.cs
@@ -6,5 +6,5 @@ public class AiModelMetadata
     public required string FriendlyName { get; init; }
     public required string Vendor { get; init; }
     public required string Description { get; init; }
-    public int? FileSizeBytes { get; set; }
+    public long? FileSizeBytes { get; set; }
 }
